// client/src/app/core/models/enums/tipo-estudio.enum.ts
export enum EnumTipoEstudioGeofisico {
    GPR = 'GPR',
    MASW = 'MASW',
    PERFILAJE_POZO = 'PERFILAJE_POZO',
    SEV_SCHLUMBERGER = 'SEV_SCHLUMBERGER',
    SEV_WENNER = 'SEV_WENNER',
    TOMOGRAFIA_ELECTRICA_DIPOLO_DIPOLO = 'TOMOGRAFIA_ELECTRICA_DIPOLO_DIPOLO',
    TOMOGRAFIA_ELECTRICA_DIPOLO_POLO = 'TOMOGRAFIA_ELECTRICA_DIPOLO_POLO',
    TOMOGRAFIA_ELECTRICA_POLO_POLO = 'TOMOGRAFIA_ELECTRICA_POLO_POLO',
    TOMOGRAFIA_SISMICA = 'TOMOGRAFIA_SISMICA'
  }
  